Revert "feat: add ibEth Asset details (#29)"
This reverts commit 24fdd3f6e931e37add82c601bf599d78e56560a4.
