+++ package.json
@@ -22,11 +22,11 @@
     "@enzymefinance/protocol": "^1.0.0-dev.5049+a3e3daba",
     "@graphprotocol/graph-cli": "^0.20.0",
     "@graphprotocol/graph-ts": "^0.20.0",
-    "@openzeppelin/contracts": "^3.4.1",
+    "@openzeppelin/contracts": "^3.4.0",
     "@types/glob": "^7.1.3",
-    "@types/node": "^14.14.33",
+    "@types/node": "^14.14.31",
     "@types/yargs": "^16.0.0",
-    "assemblyscript": "^0.18.14",
+    "assemblyscript": "^0.18.11",
     "glob": "^7.1.6",
     "graphql-request": "^3.4.0",
     "handlebars": "^4.7.7",
@@ -34,7 +34,7 @@
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
     "ts-node": "^9.1.1",
-    "typescript": "^4.2.3",
+    "typescript": "^4.2.2",
     "yargs": "^16.2.0"
   }
 }
