+++ src/entities/Asset.ts
@@ -66,10 +66,17 @@ export function extractAssets(ids: string[]): Asset[] {
   return assets;
 }
 
+export function checkChai(derivative: Asset): void {
+  if (derivative.symbol != 'CHAI') {
+    return;
+  }
+
+  derivative.derivativeType = 'Chai';
+  derivative.save();
+}
+
 export function checkSynthetix(derivative: Asset): void {
   // TODO: find better identifier of Synth assets
-  // TODO: observe SynthAdded event in the SynthetixPriceFeed
-
   if (!derivative.name.startsWith('Synth ')) {
     return;
   }
