+++ src/entities/UniswapV2PoolAssetDetails.ts
@@ -8,8 +8,6 @@ export function checkUniswapV2PoolAssetDetail(derivative: Asset): void {
     return;
   }
 
-  // TODO: monitor events in the UniswapV2PoolPriceFeed and get data from there
-
   let uniswapPair = IUniswapV2Pair.bind(Address.fromString(derivative.id));
 
   let details = new UniswapV2PoolAssetDetail(derivative.id);
@@ -27,6 +25,6 @@ export function checkUniswapV2PoolAssetDetail(derivative: Asset): void {
 
   derivative.name = name;
   derivative.derivativeType = 'UniswapPool';
-  derivative.uniswapV2PoolAssetDetails = details.id;
+  derivative.uniswapV2PoolAssetDetail = details.id;
   derivative.save();
 }
